refactor: rename subgraph to furo-stream, prepare to extract vesting
