+++ the-graph/subgraphs/furo-stream/README.md
@@ -1,10 +1,10 @@
-# Furo Subgraph
+# Furo Stream Subgraph
 
+## Unit testing
 
-# Unit testing
+### Run
 
-## Run
-```
+```sh
 yarn graph test <ONE-OR-MORE-TEST-NAMES>
 
 yarn graph test Stream
