+++ the-graph/subgraphs/furo-stream/abis/ERC20.json
