+++ the-graph/subgraphs/furo-stream/src/functions/token.ts
@@ -1,8 +1,8 @@
 import { Address, BigInt, ethereum } from '@graphprotocol/graph-ts'
-import { ERC20 } from '../../generated/FuroStream/ERC20'
-import { NameBytes32 } from '../../generated/FuroStream/NameBytes32'
-import { SymbolBytes32 } from '../../generated/FuroStream/SymbolBytes32'
-import { Token } from '../../generated/schema'
+import { ERC20 } from '../FuroStream/ERC20'
+import { NameBytes32 } from '../FuroStream/NameBytes32'
+import { SymbolBytes32 } from '../FuroStream/SymbolBytes32'
+import { Token } from '../schema'
 
 
 export function getOrCreateToken(id: string, event: ethereum.Event): Token {
