+++ the-graph/subgraphs/furo-stream/tests/mocks.ts
@@ -4,7 +4,7 @@ import {
   LogCancelStream as CancelStreamEvent,
   LogCreateStream as CreateStreamEvent,
   LogWithdrawFromStream as WithdrawEvent,
-} from '../generated/FuroStream/FuroStream'
+} from './FuroStream/FuroStream'
 
 export function createStreamEvent(
   streamId: BigInt,
