--- the-graph/subgraphs/furo/package.json
@@ -1,25 +1,0 @@
-{
-  "name": "@sushiswap/furo-subgraph",
-  "author": "Matthew Lilley <hello@matthewlilley.com>",
-  "license": "MIT",
-  "repository": {
-    "url": "sushiswap/subgraphs",
-    "directory": "subgraphs/furo"
-  },
-  "files": [
-    "generated"
-  ],
-  "scripts": {
-    "codegen": "graph codegen",
-    "build": "graph build",
-    "deploy": "graph deploy --node https://api.thegraph.com/deploy/ sushiswap/furo",
-    "create-local": "graph create --node http://localhost:8020/ sushiswap/furo",
-    "remove-local": "graph remove --node http://localhost:8020/ sushiswap/furo",
-    "deploy-local": "graph deploy --node http://localhost:8020/ --ipfs http://localhost:5001 sushiswap/furo"
-  },
-  "devDependencies": {
-    "@graphprotocol/graph-cli": "^0.27.0",
-    "@graphprotocol/graph-ts": "^0.24.1",
-    "matchstick-as": "^0.4.2"
-  }
-}
